🚚 refactor: entities/project 신설, 프로젝트 도메인 데이터 추출
[변경 내용]
- entities/project 신설: api/matchingProject(355줄), model/matchingProject 및
  각각의 test/mock 파일 이동. ProjectStatus 등 프로젝트 도메인 타입이 여기 포함
- project/list 내부의 matchingProject 상대 import 16개 파일을 절대 경로로 재작성
- 외부 feature(matching, application, settings)의 matchingProject import 경로 갱신

[변경 이유]
- matchingProject(프로젝트 목록/상세/모집 데이터)는 matching, application,
  settings가 공유하는 프로젝트 도메인 데이터라 entities가 올바른 위치
- ProjectStatus 타입을 13개 파일이 참조 중인데 features/project에 있어 cross-feature
  참조를 유발. entities로 옮겨 features -> entities 정상 방향으로 전환
- boundaries 위반 83 -> 77건 (-6): matching->project 3->1,
  application->project 11->7 (데이터 참조분 해소)

[고려 사항]
- test/mock 파일도 엔티티와 함께 이동: 프로젝트 데이터를 테스트/목킹하므로 데이터와
  콜로케이트하는 것이 응집도 측면에서 맞음
- matchingProject.mock이 projectFilterOptions(feature) 경유로 SCHOOLS_BY_BRANCH를
  참조하던 것을 원본 @/shared/config/schools 직접 참조로 변경 -> entity->feature
  역참조 방지 (projectFilterOptions는 단순 re-export였음)
- matchingProjectList(목록 필터링 로직)는 엔티티가 아닌 list feature 로직이므로
  features/project/list에 잔류, entities/project를 import
- vi.mock의 상대경로 문자열은 import 자동 치환에 안 걸려 별도 수정
  (ProjectApplyModal.ui.test)
- 남은 project 관련 위반: application<->project UI 누수(SectionHeader,
  ProjectDetailCard 등)와 matching/project->application의 application 도메인
  데이터. 전자는 SectionHeader의 shared 이동으로, 후자는 entities/application
  추출로 후속 해소 예정
- ProjectApplyModal.ui.test 5건 실패는 기존 이슈로 무관
