+++ src/features/matching/model/matchingStatusMapper.ts
@@ -2,8 +2,8 @@
 
 import { toRoundNumber } from "@/features/application/model/mappers"
 
+import type { ProjectMembersResponse } from "@/entities/project/api/matchingProject"
 import type { ManagedProjectSummaryResponse } from "@/features/application/model/apiTypes"
-import type { ProjectMembersResponse } from "@/features/project/list/api/matchingProject"
 import type { Part } from "@/shared/model/domain"
 
 import type {
