+++ src/features/project/list/index.ts
@@ -1,13 +1,3 @@
-export { isRecruitDone } from "./model/matchingProject"
-export type {
-  MatchingProject,
-  ProjectCoverImage,
-  ProjectRecruitRow,
-} from "./model/matchingProject"
-export {
-  DEFAULT_MATCHING_PROJECT_MOCK,
-  MOCK_MATCHING_PROJECTS,
-} from "./model/matchingProject.mock"
 export { useMatchingProjectListFilters } from "./model/matchingProjectList"
 export type {
   MatchingProjectListFilterDescriptor,
@@ -19,3 +9,13 @@ export { MatchingProjectCard } from "./ui/MatchingProjectCard"
 export type { MatchingProjectCardVariant } from "./ui/MatchingProjectCard"
 export { MatchingProjectsListPage } from "./ui/MatchingProjectsListPage"
 export { ProjectSearchField } from "./ui/ProjectSearchField"
+export { isRecruitDone } from "@/entities/project/model/matchingProject"
+export type {
+  MatchingProject,
+  ProjectCoverImage,
+  ProjectRecruitRow,
+} from "@/entities/project/model/matchingProject"
+export {
+  DEFAULT_MATCHING_PROJECT_MOCK,
+  MOCK_MATCHING_PROJECTS,
+} from "@/entities/project/model/matchingProject.mock"
