+++ src/features/project/list/model/matchingProjectList.ts
@@ -5,18 +5,24 @@ import { useCallback, useEffect, useMemo, useState } from "react"
 import { isOperator } from "@/entities/member/model/identity"
 import { useViewerIdentity } from "@/entities/member/view-mode/useViewerIdentity"
 import { getChaptersWithSchools } from "@/entities/organization/api/organization"
+import {
+  getMatchingProjects,
+  type ProjectItem,
+} from "@/entities/project/api/matchingProject"
 import { projectKeys } from "@/features/project/new/api"
 import { useActiveGisu } from "@/shared/hooks/useActiveGisu"
 import { formatSchoolName } from "@/shared/lib/formatSchoolName"
 
-import { getMatchingProjects, type ProjectItem } from "../api/matchingProject"
 import {
   PART_OPTIONS,
   type ProjectFilterOption,
   RECRUIT_STATUS_OPTIONS,
 } from "./projectFilterOptions"
 
-import type { PartQuotaStatus, ProjectPart } from "../api/matchingProject"
+import type {
+  PartQuotaStatus,
+  ProjectPart,
+} from "@/entities/project/api/matchingProject"
 
 export const MATCHING_PROJECT_PAGE_SIZE = 15
