+++ src/features/project/list/ui/apply-modal/ProjectApplyModal.tsx
@@ -13,6 +13,14 @@ import { Controller, useForm } from "react-hook-form"
 
 import { useToastStore } from "@/components/toast/useToastStore"
 import { useResourcePermission } from "@/entities/member/hooks/useResourcePermission"
+import {
+  createApplicationDraft,
+  getApplicationDetail,
+  getMyApplications,
+  saveApplicationDraft,
+  submitApplication,
+} from "@/entities/project/api/matchingProject"
+import { isRecruitDone } from "@/entities/project/model/matchingProject"
 import { uploadFileFlow } from "@/features/project/new/api/storage"
 import { trackEvent } from "@/shared/analytics"
 import CheckIcon from "@/shared/assets/icon/check/CheckIcon"
@@ -34,13 +42,6 @@ import { QuestionItemTitle } from "@/shared/ui/question-field/QuestionItemTitle"
 import { TextQuestionField } from "@/shared/ui/question-field/TextQuestionField"
 
 import {
-  createApplicationDraft,
-  getApplicationDetail,
-  getMyApplications,
-  saveApplicationDraft,
-  submitApplication,
-} from "../../api/matchingProject"
-import {
   type ApplyAnswerValue,
   buildAnswerPayload,
   getOptionValue,
@@ -53,20 +54,18 @@ import {
   defaultByFieldType,
   type UploadedFileValue,
 } from "../../model/applyValidation"
-import { isRecruitDone } from "../../model/matchingProject"
 import { selectCurrentApplicationForProject } from "../../model/projectDetailCta"
 import { ApplyFormSkeleton } from "./ApplyFormSkeleton"
 import { ApplyProjectTitleCard } from "./ApplyProjectTitleCard"
 
 import type { FieldErrors, Resolver } from "react-hook-form"
 
+import type { MyProjectApplicationResponse } from "@/entities/project/api/matchingProject"
+import type { MatchingProject } from "@/entities/project/model/matchingProject"
 import type {
   Question,
   Section,
 } from "@/features/project/new/model/applicationQuestion"
-
-import type { MyProjectApplicationResponse } from "../../api/matchingProject"
-import type { MatchingProject } from "../../model/matchingProject"
 
 const FILE_UPLOAD_CATEGORY = "POST_ATTACHMENT" as const
 const PORTFOLIO_UPLOAD_CATEGORY = "PORTFOLIO" as const
