+++ src/features/project/list/ui/apply-modal/RecruitQuestionsViewModal.tsx
@@ -1,5 +1,6 @@
 import { useMemo } from "react"
 
+import { isRecruitDone } from "@/entities/project/model/matchingProject"
 import { RecruitStatusChip } from "@/shared/ui/chip/RecruitStatusChip"
 import { FormHeader } from "@/shared/ui/FormHeader"
 import { CheckboxList } from "@/shared/ui/input/checkbox/CheckboxList"
@@ -10,15 +11,13 @@ import { PortfolioField } from "@/shared/ui/question-field/PortfolioField"
 import { QuestionItemTitle } from "@/shared/ui/question-field/QuestionItemTitle"
 import { TextQuestionField } from "@/shared/ui/question-field/TextQuestionField"
 
-import { isRecruitDone } from "../../model/matchingProject"
 import { ApplyProjectTitleCard } from "./ApplyProjectTitleCard"
 
+import type { MatchingProject } from "@/entities/project/model/matchingProject"
 import type {
   Question,
   Section,
 } from "@/features/project/new/model/applicationQuestion"
-
-import type { MatchingProject } from "../../model/matchingProject"
 
 const OPTION_LIST_CLASS =
   "border-teal-gray-150 flex flex-col gap-0.5 rounded-[12px] border bg-[color-mix(in_srgb,var(--color-teal-50)_40%,white)] p-1"
